+++ org.hl7.fhir.validation.cli/src/main/java/org/hl7/fhir/validation/cli/param/parsers/ValidationEngineParametersParser.java
@@ -1,14 +1,10 @@
 package org.hl7.fhir.validation.cli.param.parsers;
 
-import org.hl7.fhir.exceptions.FHIRException;
-import org.hl7.fhir.r5.terminologies.JurisdictionUtilities;
-import org.hl7.fhir.r5.utils.validation.BundleValidationRule;
 import org.hl7.fhir.utilities.Utilities;
 import org.hl7.fhir.utilities.VersionUtilities;
 import org.hl7.fhir.utilities.filesystem.ManagedFileAccess;
 import org.hl7.fhir.validation.cli.param.Arg;
 import org.hl7.fhir.validation.cli.param.IParamParser;
-import org.hl7.fhir.validation.cli.param.Params;
 import org.hl7.fhir.validation.service.model.ValidationEngineParameters;
 
 import java.io.File;
@@ -24,18 +20,13 @@ public class ValidationEngineParametersParser implements IParamParser<Validation
   public static final String IMPLEMENTATION_GUIDE = "-ig";
   public static final String DEFINITION = "-defn";
   public static final String RESOLUTION_CONTEXT = "-resolution-context";
-  public static final String JURISDICTION = "-jurisdiction";
   public static final String AI_SERVICE = "-ai-service";
   public static final String CERT = "-cert";
   public static final String TERMINOLOGY = "-tx";
   public static final String TERMINOLOGY_LOG = "-txLog";
   public static final String TERMINOLOGY_CACHE = "-txCache";
   public static final String TERMINOLOGY_CACHE_CLEAR = "-clear-tx-cache";
-  public static final String CHECK_IPS_CODES = "-check-ips-codes";
-  public static final String DO_IMPLICIT_FHIRPATH_STRING_CONVERSION = "-implicit-fhirpath-string-conversions";
-  public static final String ALLOW_DOUBLE_QUOTES = "-allow-double-quotes-in-fhirpath";
   public static final String ADVISOR_FILE = "-advisor-file";
-  public static final String BUNDLE = "-bundle";
   public static final String LOCALE = "-locale";
   public static final String LANGUAGE = "-language";
   public static final String CHECK_REFERENCES = "-check-references";
@@ -43,6 +34,8 @@ public class ValidationEngineParametersParser implements IParamParser<Validation
   public static final String NO_INTERNAL_CACHING = "-no-internal-caching";
   public static final String DISABLE_DEFAULT_RESOURCE_FETCHER = "-disable-default-resource-fetcher";
   public static final String LOG = "-log";
+  public static final String DISPLAY_WARNINGS = "-display-issues-are-warnings";
+  public static final String NO_EXTENSIBLE_BINDING_WARNINGS = "-no-extensible-binding-warnings";
 
   ValidationEngineParameters validationEngineParameters = new ValidationEngineParameters();
   @Override
@@ -91,13 +84,6 @@ public class ValidationEngineParametersParser implements IParamParser<Validation
       } else if (args[i].getValue().equals(RESOLUTION_CONTEXT)) {
         validationEngineParameters.setResolutionContext(args[i + 1].getValue());
         Arg.setProcessed(args, i, 2, true);
-      } else if (args[i].getValue().equals(JURISDICTION)) {
-        if (i + 1 == args.length)
-          throw new Error("Specified -jurisdiction without indicating jurisdiction");
-        else {
-          validationEngineParameters.setJurisdiction(processJurisdiction(args[i + 1].getValue()));
-          Arg.setProcessed(args, i, 2, true);
-        }
       } else if (args[i].getValue().equals(AI_SERVICE)) {
         validationEngineParameters.setAIService(args[i + 1].getValue());
         Arg.setProcessed(args, i, 2, true);
@@ -145,15 +131,6 @@ public class ValidationEngineParametersParser implements IParamParser<Validation
       } else if (args[i].getValue().equals(TERMINOLOGY_CACHE_CLEAR)) {
         validationEngineParameters.setClearTxCache(true);
         args[i].setProcessed(true);
-      } else if (args[i].getValue().equals(CHECK_IPS_CODES)) {
-        validationEngineParameters.setCheckIPSCodes(true);
-        args[i].setProcessed(true);
-      } else if (args[i].getValue().equals(DO_IMPLICIT_FHIRPATH_STRING_CONVERSION)) {
-        validationEngineParameters.setDoImplicitFHIRPathStringConversion(true);
-        args[i].setProcessed(true);
-      } else if (args[i].getValue().equals(ALLOW_DOUBLE_QUOTES)) {
-        validationEngineParameters.setAllowDoubleQuotesInFHIRPath(true);
-        args[i].setProcessed(true);
       } else if (args[i].getValue().equals(ADVISOR_FILE)) {
         if (i + 1 == args.length)
           throw new Error("Specified -advisor-file without indicating file");
@@ -173,19 +150,6 @@ public class ValidationEngineParametersParser implements IParamParser<Validation
             validationEngineParameters.setAdvisorFile(advisorFile);
           }
           Arg.setProcessed(args, i, 2, true);
-        }
-      } else if (args[i].getValue().equals(BUNDLE)) {
-        if (i + 1 == args.length) {
-          throw new Error("Specified -bundle without indicating bundle rule");
-        } else {
-          String rule = args[i + 1].getValue();
-          if (i + 2 == args.length) {
-            throw new Error("Specified -bundle without indicating profile source");
-          } else {
-            String profile = args[i + 2].getValue();
-            validationEngineParameters.addBundleValidationRule(new BundleValidationRule().setRule(rule).setProfile(profile));
-            Arg.setProcessed(args, i, 3, true);
-          }
         }
       } else if (args[i].getValue().equals(LOCALE)) {
         if (i + 1 == args.length) {
@@ -231,19 +195,12 @@ public class ValidationEngineParametersParser implements IParamParser<Validation
           validationEngineParameters.setMapLog(args[i + 1].getValue());
           Arg.setProcessed(args, i, 2, true);
         }
-      }
-    }
-  }
-
-  private static String processJurisdiction(String s) {
-    if (s.startsWith("urn:iso:std:iso:3166#") || s.startsWith("urn:iso:std:iso:3166:-2#") || s.startsWith("http://unstats.un.org/unsd/methods/m49/m49.htm#")) {
-      return s;
-    } else {
-      String v = JurisdictionUtilities.getJurisdictionFromLocale(s);
-      if (v != null) {
-        return v;
-      } else {
-        throw new FHIRException("Unable to understand Jurisdiction '"+s+"'");
+      } else if (args[i].getValue().equals(DISPLAY_WARNINGS)) {
+        validationEngineParameters.setDisplayWarnings(true);
+        args[i].setProcessed(true);
+      } else if (args[i].getValue().equals(NO_EXTENSIBLE_BINDING_WARNINGS)) {
+        validationEngineParameters.setNoExtensibleBindingMessages(true);
+        args[i].setProcessed(true);
       }
     }
   }
